+++ src/backend/joanie/core/flows/__init__.py
@@ -1,0 +1,6 @@
+"""
+Viewflow Flows of Joanie "Core" application
+https://github.com/viewflow/viewflow
+"""
+
+from .order import OrderFlow