+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to
 
 ### Changed
 
+- Migrate from `django-fsm` to `viewflow.fsm` 
 - Use generic AdminCourseProductRelationSerializer
 - Format all displayed date to the format "10/14/1983, 1:30 PM"